weekend-chef-backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via pyfcm
aiosignal==1.4.0
    # via aiohttp
amqp==5.3.1
    # via kombu
asgiref==3.9.2
    # via
    #   -r requirements.in
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   aiohttp
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==25.4.16
    # via twisted
billiard==4.2.2
    # via celery
cachetools==5.5.2
    # via google-auth
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via cryptography
channels==4.3.1
    # via channels-redis
channels-redis==4.3.0
    # via -r requirements.in
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
constantly==23.10.4
    # via twisted
cryptography==46.0.1
    # via
    #   -r requirements.in
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via -r requirements.in
django==5.2.6
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.9.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.40.3
    # via pyfcm
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   twisted
    #   yarl
incremental==24.7.2
    # via twisted
kombu==5.5.4
    # via celery
msgpack==1.1.1
    # via channels-redis
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via kombu
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
    #   service-identity
pyasn1-modules==0.4.2
    # via
    #   google-auth
    #   service-identity
pycparser==2.23
    # via cffi
pyfcm==2.1.0
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.3.0
    # via twisted
python-dateutil==2.9.0.post0
    # via celery
pytz==2025.2
    # via -r requirements.in
redis==6.4.0
    # via
    #   -r requirements.in
    #   channels-redis
requests==2.32.5
    # via
    #   -r requirements.in
    #   pyfcm
rsa==4.9.1
    # via google-auth
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
twisted[tls]==25.5.0
    # via daphne
txaio==25.6.1
    # via autobahn
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   pyopenssl
    #   twisted
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via
    #   pyfcm
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
yarl==1.20.1
    # via aiohttp
zope-interface==8.0.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
